Refactor: create definitions factory
